Portfolio
# My Portfolio

Welcome to my GitHub repository! This is the home of my personal portfolio, where I showcase my skills, projects, and experience as a web developer and designer.

## About Me

I'm [Your Name], a passionate and detail-oriented professional in the field of web development. With expertise in front-end technologies such as HTML, CSS, and JavaScript, I strive to create engaging and responsive websites.

## Projects

### Project 1: [Project Name]
Description of Project 1. Highlight key features and technologies used.

[Link to Project 1]

### Project 2: [Project Name]
Description of Project 2. Emphasize the impact and significance of the project.

[Link to Project 2]

## Skills

- HTML5
- CSS3
- JavaScript
- Responsive Web Design


Thank you for visiting my portfolio!